+++ borisova.ekaterina/B5/point.hpp
@@ -10,7 +10,8 @@ namespace borisova
   {
     int x, y;
   };
-  double getSquareDistance(const Point& first, const Point& second);
+
+  int getSquareDistance(const Point& first, const Point& second);
   bool isPerpendicularity(const Point& one, const Point& two, const Point three);
   std::istream& operator>>(std::istream& in, Point& point);
   std::ostream& operator<<(std::ostream& out, const Point& point);
